+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { ChakraProvider } from '@chakra-ui/react';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter } from 'react-router-dom';
-import queryClient from '@/queries/queryClient';
+import queryClient from '@/lib/queryClient';
 import { worker } from '@/mocks/browser';
 import { theme } from '@/lib/styles/theme';
 import Router from './Router';
