+++ src/Router.tsx
@@ -1,12 +1,22 @@
+import { lazy, Suspense } from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
-import OrderListPage from '@/pages/OrderListPage';
 import Layout from '@/components/Layout';
+import LoadingFallback from './components/LoadingFallback';
+
+const AdminPage = lazy(() => import('@/pages/AdminPage'));
 
 const Router = () => {
   return (
     <Routes>
       <Route element={<Layout />}>
-        <Route path="/admin/order" element={<OrderListPage />} />
+        <Route
+          path="/admin/order"
+          element={
+            <Suspense fallback={<LoadingFallback />}>
+              <AdminPage />
+            </Suspense>
+          }
+        />
         <Route
           path="/*"
           element={<Navigate to="/admin/order" replace={true} />}
