--- src/interface/props.d.ts
@@ -1,5 +1,0 @@
-export interface IOnSetParams {
-  pageValue?: number;
-  dateValue?: string;
-  event?: React.ChangeEvent<HTMLInputElement>;
-}
