+++ src/lib/hooks/useGetOrderData.ts
@@ -1,10 +1,12 @@
 import { useQuery } from '@tanstack/react-query';
 import { getOrderData } from '@/api/order';
 
-export const useGetOrderData = (pageNum = 1, date: string | null) => {
+const useGetOrderData = (pageNum = 1, date: string | null) => {
   return useQuery({
     queryKey: ['/mock/order', pageNum, date],
     queryFn: () => getOrderData(pageNum - 1, date).then((res) => res.data),
-    // refetchInterval: 5000,
+    refetchInterval: 5000,
   });
 };
+
+export default useGetOrderData;
