+++ src/lib/hooks/useSetParams.ts
@@ -1,10 +1,10 @@
 import { useSearchParams } from 'react-router-dom';
-import { IOnSetParams } from '@/interface/props';
+import { IOnSetParams } from '@/interface/main';
 
 const useSetParams = () => {
   const [searchParams, setSearchParams] = useSearchParams();
-  const PAGE = Number(searchParams.get('page')) || 1;
-  const DATE = searchParams.get('date');
+  const currentPage = Number(searchParams.get('page')) || 1;
+  const currentDate = searchParams.get('date');
 
   const onSetParams = ({ pageValue, dateValue, event }: IOnSetParams) => {
     if (pageValue !== undefined) searchParams.set('page', String(pageValue));
@@ -15,7 +15,7 @@ const useSetParams = () => {
     window.scrollTo(0, 0);
   };
 
-  return { PAGE, DATE, onSetParams };
+  return { currentPage, currentDate, onSetParams };
 };
 
 export default useSetParams;
