+++ src/mocks/handlers/index.ts
@@ -1,3 +1,3 @@
-import * as orderListHandlers from './order';
+import { orderListHandlers } from './order';
 
 export const handlers = [...Object.values(orderListHandlers)];
