+++ src/pages/AdminPage.tsx
@@ -1,0 +1,15 @@
+import StatsArea from '@/components/StatsArea';
+import OrderTableArea from '@/components/OrderTableArea';
+import DatePicker from '@/components/DatePicker';
+
+const AdminPage = () => {
+  return (
+    <>
+      <StatsArea />
+      <DatePicker />
+      <OrderTableArea />
+    </>
+  );
+};
+
+export default AdminPage;
