--- src/queries/queryClient.ts
@@ -1,5 +1,0 @@
-import { QueryClient } from '@tanstack/react-query';
-
-const queryClient = new QueryClient({});
-
-export default queryClient;
